Allow escaped string filters
Let's say you have a `keywords` filter where users may put commas in the
input. Prior to this commit, we would interpret that comma as a
delimiter and treat this incorrectly as an array of multiple values.

You can now escape values to ensure the desired behavior:

```
?filter[keywords]={{foo,bar}}
```

Would end up as `.where(keywords: "foo,bar")`

This works with arrays:

```
?filter[keywords]={{foo,bar}},baz,{{another,thing}}
```

Would end up as `.where(keywords: ["foo,bar", "baz", "another,thing"])`

Originally, quoting the string was considered. However, for this use
case it would be common for users to type quotes, and then you'd have to
escape the quotes. Using double-curlies keeps things simple and without
conflict.